+++ app/src/main/java/com/example/tidy/MainActivity.kt
@@ -18,21 +18,17 @@
 package com.example.tidy
 
 import android.Manifest
-import android.app.AlarmManager
 import android.app.NotificationChannel
 import android.app.NotificationManager
 import android.content.Context
-import android.content.Intent
 import android.content.pm.PackageManager
 import android.os.Build
 import android.os.Bundle
-import android.provider.Settings
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
 import androidx.activity.result.contract.ActivityResultContracts
 import androidx.core.content.ContextCompat
-import androidx.core.net.toUri
 import androidx.core.splashscreen.SplashScreen.Companion.installSplashScreen
 import com.example.tidy.ui.screen.MainScreen
 import com.example.tidy.ui.theme.TidyTheme
@@ -60,7 +56,7 @@ class MainActivity : ComponentActivity() {
         }
         createNotificationChannel(this)
         askNotificationPermission()
-        requestExactAlarmPermission(this)
+        Utils.requestExactAlarmPermission(this)
     }
 
     private fun createNotificationChannel(context: Context) {
@@ -88,18 +84,6 @@ class MainActivity : ComponentActivity() {
                 ) != PackageManager.PERMISSION_GRANTED
             ) {
                 requestPermissionLauncher.launch(Manifest.permission.POST_NOTIFICATIONS)
-            }
-        }
-    }
-
-    private fun requestExactAlarmPermission(context: Context) {
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.S) {
-            val alarmManager = context.getSystemService(ALARM_SERVICE) as AlarmManager
-            if (!alarmManager.canScheduleExactAlarms()) {
-                val intent = Intent(Settings.ACTION_REQUEST_SCHEDULE_EXACT_ALARM).apply {
-                    data = "package:${context.packageName}".toUri()
-                }
-                context.startActivity(intent)
             }
         }
     }
